model: split the 1285-line module into filters, losses and blocks
model.py held the backbone wiring, the three filter heads, the conv blocks
they are built from, and the loss, so the file someone opens to understand the
method was the longest file in the repository.

The four parts are now filters.py (the three parametric filters and the
coordinate grids), losses.py (Eq. 8 and its MS-SSIM), blocks.py (the conv
blocks the heads stack), and model.py (ColourHead, the parameter prediction
head, and DeepLPFNet). Class bodies are moved unchanged and the module
hierarchy is untouched, so state-dict keys are identical and existing
checkpoints load as before; the names that used to live in model.py are
re-exported from it, so model.CubicFilter and model.DeepLPFLoss still resolve.

47 tests pass, including the pretrained-checkpoint replication test.
