+++ blocks.py
@@ -1,0 +1,123 @@
+# -*- coding: utf-8 -*-
+#Copyright (C) 2020. Huawei Technologies Co., Ltd. All rights reserved.
+
+#This program is free software; you can redistribute it and/or modify it under the terms of the BSD-3-Clause License.
+
+#This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the BSD-3-Clause License for more details.
+"""The convolutional blocks the three filter heads are built from.
+
+Each head stacks the same four blocks - a 3x3 convolution with leaky ReLU and
+dropout, max pooling, and a global average pool - so they live here rather than
+being repeated per filter. Split out of ``model.py`` unchanged; the class names
+and attribute names are the ones the released checkpoints were saved with.
+"""
+import torch.nn as nn
+
+
+class Block(nn.Module):
+    """Base class providing shared convolution helpers for the conv blocks."""
+
+    def __init__(self):
+        """Initialisation for a lower-level DeepLPF conv block
+
+        :returns: N/A
+        :rtype: N/A
+
+        """
+        super(Block, self).__init__()
+
+    def conv3x3(self, in_channels, out_channels, stride=1):
+        """Represents a convolution of shape 3x3
+
+        :param in_channels: number of input channels
+        :param out_channels: number of output channels
+        :param stride: the convolution stride
+        :returns: convolution function with the specified parameterisation
+        :rtype: function
+
+        """
+        return nn.Conv2d(in_channels, out_channels, kernel_size=3,
+                         stride=stride, padding=1, bias=True)
+
+
+class ConvBlock(Block, nn.Module):
+    """3x3 strided convolution followed by a LeakyReLU non-linearity."""
+
+    def __init__(self, num_in_channels, num_out_channels, stride=1):
+        """Initialise function for the higher level convolution block
+
+        :param num_in_channels: number of input channels
+        :param num_out_channels: number of output channels
+        :param stride: unused; the internal convolution uses a fixed stride of 2
+        :returns: N/A
+        :rtype: N/A
+
+        """
+        super(Block, self).__init__()
+        self.conv = self.conv3x3(num_in_channels, num_out_channels, stride=2)
+        self.lrelu = nn.LeakyReLU()
+
+    def forward(self, x):
+        """ Forward function for the higher level convolution block
+
+        :param x: Tensor representing the input BxCxWxH, where B is the batch size, C is the number of channels, W and H are the width and image height
+        :returns: Tensor representing the output of the block
+        :rtype: Tensor
+
+        """
+        img_out = self.lrelu(self.conv(x))
+        return img_out
+
+
+class MaxPoolBlock(Block, nn.Module):
+    """2x2 max-pooling block with stride 2 (halves the spatial resolution)."""
+
+    def __init__(self):
+        """Initialise function for the max pooling block
+
+        :returns: N/A
+        :rtype: N/A
+
+        """
+        super(Block, self).__init__()
+
+        self.max_pool = nn.MaxPool2d(kernel_size=2, stride=2)
+
+    def forward(self, x):
+        """ Forward function for the max pooling block
+
+        :param x: Tensor representing the input BxCxWxH, where B is the batch size, C is the number of channels, W and H are the width and image height
+        :returns: Tensor representing the output of the block
+        :rtype: Tensor
+
+        """
+        img_out = self.max_pool(x)
+        return img_out
+
+
+class GlobalPoolingBlock(Block, nn.Module):
+    """Global average-pooling block collapsing each feature map to a scalar."""
+
+    def __init__(self, receptive_field):
+        """Implementation of the global pooling block. Takes the average over a 2D receptive field.
+        :param receptive_field: nominal receptive field size (unused; an
+            adaptive average pool to a 1x1 output is used instead)
+        :returns: N/A
+        :rtype: N/A
+
+        """
+        super(Block, self).__init__()
+        self.avg_pool = nn.AdaptiveAvgPool2d(1)
+
+    def forward(self, x):
+        """Forward function for the high-level global pooling block
+
+        :param x: Tensor of shape BxCxAxA
+        :returns: Tensor of shape BxCx1x1, where B is the batch size
+        :rtype: Tensor
+
+        """
+        out = self.avg_pool(x)
+        return out
+
+
